+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-05-10 12:09:19
-// Commit: feat(ui): add async ui stream provider
+// Updated at: 2026-05-10 17:30:09
+// Commit: feat(ui): implement shimmer loading skeleton in ui
 
 class uiComponentState {
-  final String commitRef = '994';
-  final String lastMessage = 'feat(ui): add async ui stream provider';
-  final DateTime timestamp = DateTime.parse('2026-05-10T12:09:19+05:00');
+  final String commitRef = '1014';
+  final String lastMessage = 'feat(ui): implement shimmer loading skeleton in ui';
+  final DateTime timestamp = DateTime.parse('2026-05-10T17:30:09+05:00');
 }